+++ resources/views/components/ui/sidebar.blade.php
@@ -8,7 +8,7 @@
         ['analytics', 'Analytics', route('analytics'), true, 'M4 20V10M10 20V4M16 20v-7M22 20H2'],
         ['customers', 'Customers', route('customers'), false, 'M9 8a3 3 0 100-6 3 3 0 000 6zM3 20c0-3.3 2.7-5 6-5s6 1.7 6 5'],
         ['service', 'Service', route('service'), false, 'M14 6a4 4 0 005 5l-8 8-3-3 6-6a4 4 0 010-4z'],
-        ['promo', 'Promo', route('promo'), true, 'M3 12l8-8h8v8l-8 8-8-8zM15 9h.01'],
+        ['promo', 'Promo', route('promo.index'), true, 'M3 12l8-8h8v8l-8 8-8-8zM15 9h.01'],
         ['pegawai', 'Akun Pegawai', route('pegawai'), true, 'M3 5h18v14H3zM9 12a2 2 0 100-4 2 2 0 000 4z'],
     ];
 @endphp